CMake FindX alterations
arking all as advanced without doing an if on PACKAGE_FOUND and nullifying the include_dir and library if not found will hide the entry in cmake gui on windows. Adding an entry and pointing to the correct path at this point will not work so you are forced to do a hard reset of the cache, so you have to enter every single entry exactly right before selecting 'configure' or it will fail and you'll have to do a hard reload of the cache and start from scratch, which can become very annoying.

With these changes you can open the project in CMake and press configure right away and then simply start replacing the NOTFOUND paths with the actual paths and simply press 'configure' again, no need for any hard reload of cache.

Also there's no need to hide BOTAN_ROOT_DIR and MYSQLCCPP_ROOT_DIR at all, just like PCRE_INCLUDE_DIR and PCRE_LIBRARY shouldn't be hidden either.