Shrink buffer offset types where possible
ALlow the buffer's offset types to shrink to the smallest type required to accomodate for the size of the storage block. For the current network buffer size, this saves 12 bytes per client. Think of the caches!

Could potentially shrink the pointers too but it's probably not worth the hassle to save 4-8 bytes.