Fix a bug where the user's salt started with a zero byte
The conversion to BigInt was dropping the leading zero byte(s) and causing the hash to be computed incorrectly.

Rather than encoding BigInt with encode_1363 to pad the output, treat the salt as a binary array instead. This avoids hardcoding the salt length with assumptions about the game client or needing to pass it around.

Closes https://github.com/EmberEmu/Ember/issues/27