Remove double empty buffer check for average case
process...() is only called when the buffer has data in it, so doing the empty()check at the end rather once at the start and again at the end will (typically) remove a single empty() for each call. If the function is somehow called with an empty buffer, no harm is done, it'll fail the header check. I can't spot something like this and not change it, even if it's pretty inconsequential.