Update Realmlist
Atomically load shared_ptr instead of declaring atomically.

This is to satisfy clang/libc++

It's a slight regression in optimization but an improvement in compatibility.

The operation should still be thread-safe, just one that supports a broader range of systems. Atomic_load ensures it's still an atomic operation even if it's 'emulated' and not as pretty when loading.